refactor: js 모듈 분리
