+++ src/header/handleTitleIcon.js
@@ -1,0 +1,8 @@
+const handleClickTitleIcon = () => {
+    const titleIcon = document.querySelector(".title-icon");
+    titleIcon.addEventListener("click", () => {
+        location.reload();
+    });
+};
+
+export { handleClickTitleIcon };
