+++ src/news-bar/rolling.js
